Don't check stride by default
We now have three out of three datapoints that `check_stride` will be `partial`'ed to `False`:

- `torch`: https://github.com/pytorch/pytorch/pull/58981#discussion_r639514081
- `torchvision`: https://github.com/pytorch/pytorch/issues/56544#issuecomment-845352605
- `kornia`: https://github.com/kornia/kornia/blob/9041c42b410e6a4bbb664c7134a120be80aa2265/test/utils.py#L25

Given that the strides in most cases are in implementation detail, IMO we should change the default to `False`. In cases were matching strides is a requirement for closeness / equality it can always set to `True` manually.

ghstack-source-id: f25adbac64c52dd758b57def1c527fedd7c6351a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60637