[package] ensure we don't write files twice to the archive.
The ZIP format allows for writing multiple files with the same name. But
this is handled poorly by most tooling (including our own), so doing so
produces weird behavior depending on the implementation of the ZIP
reader.

Since we have no valid use case for writing multiple files with the same
name to a `torch.package`, just ban it.

ghstack-source-id: 1a2cf47996cf831545be3b679eec1fc24982797d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61371