Avoid using mp.Manager to report #GPUs needed in dist tests
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61409

We used a multiprocessing.Manager in order to share TEST_SKIPS between the parent and the child processes. TEST_SKIPS is a global variable that defines a unique error code for each "error type", so that the parent can figure out the reason a child exited. While originally this mapping was immutable, at some point we allowed children to modify the parent's value of that mapping so they could update the message for the `multi-gpu` error to make it reflect how many GPUs were really needed. This occurred in D23285790. Since then this Manager proved to be quite problematic, especially around thread safety, races, TSAN, ... (see D22753459, D23641618, D28490129, D28794321 and D29585862). This seems like an awful lot of trouble for such a small functionality. Here I propose we drop Manager and instead get the same result by using separate error codes for each number of GPUs. It should be much simpler and thus more robust.
ghstack-source-id: 133236447

Differential Revision: [D29612614](https://our.internmc.facebook.com/intern/diff/D29612614/)