[package] minor cleanups to internal APIs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61428


I was reading this code again after a while and didn't understand as
quickly as I would have liked. Some of the function names are no longer
accurate, etc.

This PR renames these functions to be in the same language of
"dependencies" that the rest of the API uses. I think the resulting
usage of the APIs is more clear than before

Differential Revision: [D29620946](https://our.internmc.facebook.com/intern/diff/D29620946/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D29620946/)!
ghstack-source-id: ee6a2d29eeff6e9481c94f02803925be67923fb0