[package] properly handle case where we are re-packaging mocked modules
Mocking is the only time we introduce a "special" module to a
torch.package of our own creation. This interacts poorly with
re-packaging, since if we treat `_mock` as a regular module and try to
package it normally we will produce a broken package.

This PR teaches PackageExporter to recognize `_mock` modules and treat
them specially during the dependency walking process, thus avoiding the
issue.

ghstack-source-id: ff79a16218a105408a56550802b3c51372d53d5d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61434